WARN_UNUSED_RESULT: change SAL version test
Previously we tested for _Check_return_ based on the version of MSVC,
but that may not work properly if <sal.h> is not included. Furthermore,
it's possible for _Check_return_ to be defined on other platforms;
see https://github.com/nemequ/salieri for one example.

This version simply checks if _Check_return_ has been defined and, if
so, uses it.