ASSUME_ALIGNED: various enchancements
First, I added an OpenMP 4+ version (#pragma omp simd...), which is the
preferred option.

The GCC version has been changed to assign the result of the call to
__builtin_assume_aligned to the value it was passed.  This is due to
the fact that, unlike other compilers, GCC doesn't actually mark the
argument as aligned, but instead returns a value which is aligned.
IMHO GCC's version makes more sense, but unfortunately it creates a bit
of a problem for portable code.

The MSVC version has been changed to avoid pointer arithmetic, though I
don't think it was a problem in practice.

Finally, there is now also a version which uses __builtin_unreachable()
with an if statement, which may be helpful for versions of clang which
don't support __builtin_assume_alinged.