Don't trust __GNUC{_MAJOR,_PATCHLEVEL}__ for Intel.
They just set the values to whatever version of GCC is installed on
the system, which means we may have functionality available which
exceeds the GNUC macros or, more importantly, the macros may lead us
to believe that functionality is available which is not implemented by
ICC.

For example, ICC 18.0 doesn't support the fallthrough attribute, but if
you install it on a system with GCC 7 (such as my Fedora 27 machine)
__GNUC__ will be defined as 7, and GCC 7 *does* support fallthrough.

Instead of relying on the GCC compatibility, we need to test the ICC
version. Unfortunately I'm not aware of a way to tell which version of
GCC is supported by ICC before ICC 16, so a lot of functionality will
simply be disabled on older versions of ICC until someone can test
them properly (and, AFAICT, Intel doesn't make older versions
available).