Add Cray Compiler support.
The only thing I could find to implement in Hedley is
`#pragma _CRI message ...`.

There is something analogous to __builtin_expect
(#pragma _CRI probability...), but it's different enough that we can't
implement it in Hedley.  Same goes for the inlining pragmas.

Still, the version check could be useful for outside code.