Restructure version macros, add HEDLEY_*_VERSION
The new HEDLEY_*_VERSION macros establish a single way to determine
the compiler and version instead of having to look up the proper
predefined macros for each compiler. For each compiler we define a
HEDLEY_*_VERSION macro if that compiler is currently in use, with
the value set to the encoded (with HEDLEY_VERSION_ENCODE) version.

These macros are then used throughout the rest of Hedley fairly
uniformly.

This also adds HEDLEY_GNUC_* macros in addition to the HEDLEY_GCC_*
macros. For the GCC versions, the goal is to have them set for GCC,
but not other compilers which set __GNUC__ and friends, whereas the
GNUC versions will just blindly obey __GNUC__ and friends. This is
useful because various compilers will often masquerade as a version
of GCC for which they do not support all features.

Note that the GCC macros work by blacklisting compilers which are
known to set __GNUC__ (at least in some configurations). Unfortunately,
compilers which set __GNUC__ but are not know to Hedley will result
in a false positive. If you run into such a compiler please let us
know so we can add it to the list.