Add HEDLEY_DIAGNOSTIC_DISABLE_UNKNOWN_ATTRIBUTE macro
This attempts to disable warnings about C++ attributes which the
compiler doesn't know about.  For example, clang::fallthrough on
MSVC.

Note that it does not handle attributes the compiler knows about but
require a newer version of the standard (e.g., trying to use the
C++17 fallthrough attribute in C++11 mode).