Add HEDLEY_DIAGNOSTIC_DISABLE_CPP98_COMPAT_WRAP_ macro.
This is to work around probably the dumbest warning I have seen.  I'm
usually a fan of very aggressive warnings, but clag's -Wc++98-compat
is just stupid.  The idea is that you can compile your code in C++11+
mode, but get warnings whenever you accidentally use features that
C++98 doesn't support so you can maintain support for C++98 compilers.

The problem is that __cplusplus is defined to > 199711L, so code that
takes advantage of C++11+ features only after checking to see if the
compiler is in C++11+ mode will still emit warnings.  Hedley does this
a lot, so now we have to enclose any features not available in C++98
in blocks that look like

  _Pragma("clang diagnostic push")
  _Pragma("clang diagnostic ignored \"-Wc++98-compat\"")
  CODE GOES HERE
  _Pragma("clang diagnostic pop")

This macro just makes that a bit less painful.

This warning should really not exist... if people want to make sure
their code compiles in C++98 mode they can just (try to) compile in
C++98 mode.  However, since it does exist and we want to avoid any
warnings even when people compile with -Weverything, here we are.