Rename HEDLEY__IS_CONSTEXPR to HEDLEY_IS_CONSTEXPR_
This has always been an internal macro, so it's allowed.

This is mainly for consistency with the new
HEDLEY_DIAGNOSTIC_DISABLE_CPP98_COMPAT_WRAP_ macro, which I would have
liked to call HEDLEY__DIAGNOSTIC_DISABLE_CPP98_COMPAT_WRAP but clang
in C++ mode now generates a -Wreserved-id-macro warning because of the
double-underscore.  AFAIK only macros which *start with* a
double-underscore are reserved, but clang still complains and I think
this is a less-bad work-around than temporarily disabling the warning
to declare the macros.