Add HEDLEY_NULL macro.
Let me first say THIS SHOULD NOT BE USED EVERYWHERE.  I'm not trying
to redefine everything in the C/C++ standards.  I used to be pretty
heavily involved in GNOME, and I remember all the criticism about
gfloat, gint, gdouble, etc., so maybe I'm just a bit sensitive about
it, but let me try to exlain the reason a bit, and encourage you *not*
to use it without a good reason.

Basically, this is designed to switch between NULL and, in C++11+,
nullptr.  nullptr was introduced into C++11 for legitamate reasons but
it does bring a small hiccup for code that wants to be portable to
different versions of C++, and/or C.

So, if you don't care about portability to C or version of C++ before
C++11, DON'T USE THIS, use nullptr.  If you don't care about
portability to C++11+, DON'T USE THIS, use NULL.

Even if you do care about portability, it may be easier to just
understand the pitfalls of NULL and the reason nullptr was introduced
and just be careful, but you should be aware of clang's
-Wzero-as-null-pointer-constant diagnostic.  It's not enabled by
default (unless you use -Weverything), but if you want to avoid
warnings even with -Weverything then this macro may be a good way to
go.