Integrate UNREACHABLE and ASSUME tests more tightly.
The HEDLEY_UNREACHABLE, HEDLEY_UNREACHABLE_RETURN, and HEDLEY_ASSUME
macros are pretty closely related, but before now the tests were
mostly independent.  This version should be better at choosing the
right implementation and reduce duplication in the checks.

I've also done away with the possibility of HEDLEY_UNREACHABLE or
HEDLEY_ASSUME being defined to nothing; worst case now is you'll get
something like `((void) (expr))` or `((void) (0))`

Finally, I got rid of the possibility of implementing
HEDLEY_UNREACHABLE with abort().  This is meant to be used as an
optimization hint for the compiler, so I don't want to be injecting
a branch every time it's used... that could be very bad for
performance.