Add support for PGI (aka NVCC) 20.7.
Most of these changes are due to the diagnostic numbers changing by
one; I'm not sure whether that is temporary or will be the new
normal. See [1] for more information.

The other change is a bug ([2]) in __builtin_expect_with_probability
which will hopefully be fixed in the next release.

[1] https://forums.developer.nvidia.com/t/many-all-diagnostic-numbers-increased-by-1-from-previous-values/146268
[2] https://forums.developer.nvidia.com/t/compilers-claim-support-for-builtin-but-compilation-fails/146242