Disable both "new" and "old" diagnostic numbers on PGI.
It looks like the "new" diagnostic numbers (which are one greater than
the previous version) were not intentional, and I expect them to go
back to their original values in the next release.  Once they do I'll
change the checks from >= 2020.7 to == 2020.7, but I want to include
both diagnostic numbers until then so that Hedley v14 will continue to
work as expected with > 2020.7.  The disadvantage is that we disable
some warnings which we do actually want, but I think that's the lesser
of two evils here.