Don't output flag_enum attribute in C++ until clang 5.
Prior to clang 5, __has_attribute(flag_enum) would return true in both
C and C++ mode, but in C++ mode attempting to actually use it would
trigger a diagnostic about the flag_enum attribute being ignored
(-Wignored-attributes).  See the commit message from when flag_enum
was added to LLVM:

  https://github.com/llvm/llvm-project/commit/724f14e75c50c6839408509b2fede5464fa9111e

I don't have the exact commit where support was added in C++ mode, but
4.0.1 emits the diagnostic and 5.0 doesn't:

  https://godbolt.org/z/8vjjKq

Since we can't rely on the value of __clang__ or
__has_attribute(flag_enum), we try to detect clang version 5 by testing
for the -Wbitfield-enum-conversion warning, which first appeared in
clang 5.