Add support for MCST's LCC compiler.
The compiler uses the __LCC__ predefined macro, which is also used by
the other, more well-known LCC compiler (<https://drh.github.io/lcc/>).
To disambiguate we also check if __LCC_MINOR__ is defined, which MCST's
compiler does define but LCC doesn't.

Overall, the compiler's support for GNU C is pretty impressive.  A
number of attributes are silently ignored because the diagnostics they
support aren't provided by MCST LCC, but that's fine; the code is still
accepted.  In these cases we go ahead and output the attribute since
the relevant diagnostic may be added in the future.  The main
limitations seem to come from the EDG front-end; things like #pragma
warning aren't supported, nor is pushing/popping the warning stack, and
of course the diagnostics are numbered and use diag_suppress to disable
instead of GCC diagnostic ignored.

The only real problem I saw was something ICC also does, where
_Noreturn emits a diagnostic if it comes before the storage class
specifier (i.e., "static").  This is bad because C++ attributes have
to come first, so supporting both C and C++ is hard.  The workaround
is to prefer __attribute__((__noreturn__)) even in C11+ mode.