Several fixes for IAR > 8.3.
IAR has apparently started to make use of feature test macros such as
__has_attribute and __has_extension.  Unfortunately, the results appear
to be untrustworthy as several are inaccurate, at least for some
versions before 8.5.9, so until 8.5.9 we ignore __has_attribute on IAR.
To lessen the impact, I've gone through the complete list of GCC-style
attributes IAR claims to support (there are only 18) and added an
explicit check to enable them.

For __has_extension, so far only attribute_deprecated_with_message is
wrong (possibly others, but none we test), and still is as of 8.5.9, so
we currently *do* make use of __has_extension but ignore the result for
purposes of HEDLEY_DEPRECATED.  This may change at a later date if we
identify more lies.

This also silences some warnings from IAR; one about `(void) foo` being
a C-style cast, the other about `char const **` not being a standard
type for argv (the const being the problem).

Finally, some very old versions would have used VER instead of __VER__
to check the version number, which was wrong.

Fixes #46.